Use new `symbol-overload(_, _)` attribute (#984)
Following https://github.com/runtimeverification/k/pull/3989, the KORE
attribute that encodes "this axiom declares an overload equality over
two symbols `X` and `Y`" is being gradually renamed from `overload(_,_)`
to `symbol-overload(_,_)`.

This PR moves the LLVM backend over to accepting the new attribute
rather than the old one; doing so is a simple renaming process on the
attribute following #983. The checked-in KORE test files are updated by
a simple renaming of the attribute.[^1]

~~Blocked on: #983~~

[^1]: See #987; it's a pain to regenerate all the KORE at the moment so
I've edited the files manually using `sed` rather than the actual K
frontend. This is fine for this case in the frontend, but in the future
if we make more complex changes to the structure of compiled KORE
definitions we should implement a more robust way of rebuilding the test
suite to pin against a particular frontend version.