Remove usages of K error handling infrastructure (#1001)
Part of: https://github.com/runtimeverification/llvm-backend/issues/999

We are currently working on removing the reverse dependency that the
LLVM backend has on the K frontend's data structures; this PR takes a
first step towards that goal by removing the backend's use of the K
exception infrastructure for throwing errors.

The K infrastructure is replaced with a simple exception class that the
frontend will be able to catch and reinterpret as appropriate.

This PR needs a matching change to the K frontend to work, along with
appropriate lockstep merging.

---------

Co-authored-by: Tamás Tóth <tothtamas28@users.noreply.github.com>