Remove usages of K source and location classes (#1002)
Part of: https://github.com/runtimeverification/llvm-backend/issues/999

We are currently working on removing the reverse dependency that the
LLVM backend has on the K frontend's data structures; this PR takes a
step towards that goal by removing the backend's use of the K Source and
Location classes. These classes are only ever used in the matching
compiler as simple PODs that carry around data, so they can be replaced
trivially with Java classes here.

This PR needs a matching change to the K frontend to work, along with
appropriate lockstep merging.

Also contains a minor fix in
https://github.com/runtimeverification/llvm-backend/pull/1002/commits/4b644b04278cf030580276288785307167276903
for an edge case in the exhaustiveness checker for rules with no
location; we want to propagate an empty optional here rather than
`null`.