Remove usages of K APIs for KORE -> K
The backend relied on the K Frontend's libraries to perform
pretty-printing of KORE terms back to KAST when errors occur (e.g. to
print the counterexample for a non-exhaustive match). We need to remove
these usages so that the backend only depends on `scala-kore`.

In a subsequent change, we'll update the error-handling mechanism so
that the KORE term can be passed back to the frontend and pretty-printed
there.