Backport #923
In https://github.com/runtimeverification/llvm-backend/pull/923, we
removed the old Maven / Nix infrastructure but subsequently had to
revert that change because the KORE snapshot JARs from the frontend
would break the Nix hash on every K release.

This commit reinstates that change and brings the LLVM backend's Maven /
Nix code up to date with K's.

Co-authored-by: Samuel Balco <goodlyrottenapple@gmail.com>