Attach counterexample pattern to exceptions
Previously, this pattern was pretty-printed directly by the matching
compiler. Now that we no longer depend on the K Frontend, we can't do
this pretty-printing here and instead need to attach the KORE term to
the exception and bubble it up to the caller.

This will require a change in the K Frontend at the point where errors
are translated to turn this pattern into a KAST term that can be shown
to the user.