Remove usages of K frontend code (#1007)
This PR finalises the removal of the LLVM backend's dependence on the K
frontend, now that we have the shared data structures extracted out to
the [`scala-kore`
repository](https://github.com/runtimeverification/scala-kore). There
are a few smaller changes that make up this PR, but the commit history
is clean and can be reviewed commit-by-commit.

Once this is merged, the LLVM backend will no longer depend on the K
Frontend and we will have eliminated the backwards edge in the K
dependency graph!

Fixes: https://github.com/runtimeverification/llvm-backend/issues/999

---------

Co-authored-by: Samuel Balco <goodlyrottenapple@gmail.com>
Co-authored-by: rv-jenkins <admin@runtimeverification.com>