Revert "Revert "Set LLVM backend up to receive automatic updates from scala-kore"" (#1012)
Reverts runtimeverification/llvm-backend#1010

This PR reinstates the fixed Nix infrastructure that was reverted for
POM incompatibility reasons; once this is merged we will stop
experiencing the Nix hash invalidation issues that were bothering us
over the weekend. I have verified locally that the changes in
https://github.com/runtimeverification/llvm-backend/pull/1014 would have
caught the problem that we needed to revert in the first place, but this
PR needs to get merged first so that the CI will start working for that
PR.

The only non-revert changes here are to add an explicit plugin
dependency on the maven compiler with the appropriate source version set
explicitly, and to update the Nix hash as a result. The remainder of the
PR has been reviewed already.