use move assignment in store_rangemap_for_gc
the code in store_*_for_gc was assuming that the map/list/set/rangemap
types were all POD types. However, only map/list/set are in fact POD
types. Therefore, for rangemap, we use the move assignment operator in order to
ensure we copy the underlying bits rather than invoking the copy assignment
operator. This should be safe because the source memory is about to be
deallocated.