Fix `kompile -O3` (#1009)
There exists a bug in `kompile -O3` which occurs if a specialized step
function has a residual of collection sort (i.e., its rhs has a function
that returns a collection) and garbage collection is triggered
immediately after this rule applies. In this case, the collection in
question has been allocated with koreAllocAlwaysGC, and the memory it is
using gets freed. As a result, subsequent allocations can write to that
memory and corrupt the term pointed to by that collection pointer.

The fix is to detect when this occurs just prior to invoking the garbage
collector, and copying the collection pointer into a block with a block
header, which then gets passed as the updated garbage collection root to
the gc. GC keeps that memory alive, and then the post-gc address seen by
the function that invoked the collector corresponds to the migrated
pointer on the main kore heap.

We implement this with the following steps:

1. Create a family of functions `store_<cat>_for_gc` and
`load_<cat>_for_gc` which handle the process of storing and loading the
collection pointer to the heap when the GC is invoked.
2. Create new reserved symbols used to store these values for all
collection sort categories present in the definition.
3. Ensure that the garbage collector knows that a root is a block
containing a collection rather than the collection pointer itself.

---------

Co-authored-by: Bruce Collie <brucecollie82@gmail.com>