Updates to hook and function event hints (#1020)
This PR implements two updates to the format of hook and function event
hints.

1. The PR fixes an inconsistency in the hint format for hook hints,
where the hooks that are hard-coded in the code generator were including
their arguments as part of the hook hint, while the hooks that call an
external function were not including their arguments as part of the
hint. We know include the hook arguments in both cases.
2. The PR removes the arguments from function events by default. This is
done for performance reasons and given that the math proof team does not
need these arguments for their proof automation. We add an additional
flag to `llvm-kompile`, namely `--proof-hint-instrumentation-slow`, that
if passed (instead of `--proof-hint-instrumentation`) will instrument
the code in the previous way, i.e. with generation of arguments for
function event hints.