change loc of forwarding ptr
Previously the forwarding ptr was always 8 bytes after the block header.
However, in the event that the first child of a block is a map/list/set,
this causes the forwarding ptr to overwrite the base pointer offset,
which means that if we later try to locate the base pointer of a derived
pointer from that same address, we will end up with a corrupted address
for the base pointer.

The solution is pretty straightforward; previously, 8 bytes after the
block header was always the same as the offset of the first child of the
block. Thus, we make it so that we instead maintain the invariant that
the forwarding pointer is the same address as the address of the first
child of the block, which in some cases will be 16 bytes after the
block header.