Omitting intemedaite configuration events in the proof trace (#1025)
This PR disables the generation of intermediate configuration events in
the proof trace by default. It is a followup of #1020 and aims to
further improve the performance of generating hints. Similarly to #1020,
we allow the user to get the old behavior of emitting every intermediate
configuration event by passing the `--proof-hint-instrumentation-slow`
flag to `llvm-kompile`.

---------

Co-authored-by: Dwight Guth <dwight.guth@gmail.com>