Remove `rawTerm` from hints output (#1027)
Closes https://github.com/Pi-Squared-Inc/pi2/issues/1300: 

`rawTerm` is a useful wrapper on a KItem injection. However, The Math
Proof Generation Team from $Piˆ2$ doesn't use it in their parser, so
they need to unwrap the injection and drop the `rawTerm` string every
time. To avoid this unnecessary overhead, we introduce in this PR a
modification to `proof_event::emit_serialize_term` and
`serialize_term_to_file` to allow us to construct the injection without
using the `rawTerm` wrapper.

The idea of including this modification into `serialize_term_to_file`
was to avoid a third function (`serialize_k_term_inj_to_file`) that has
the exact same structure as `serialize_term_to_file` and
`serialize_raw_term_to_file` with a single different line.