Fix locations in proof hints (#1031)
This PR fixes 2 issues with the generation of locations for the hook and
function proof hint events.

- The starting location stack is now the empty stack as opposed to the
`"0"` stack. This is because the math team assumes that the first
location in the stack corresponds to an argument of the top cell in the
configuration and not the top cell constructor.
- The previous method for skipping locations that correspond to
injections, namely skipping a location when the child of a term is an
injection, was leading to errors when it combined with the injection
compaction optimization. Specifically, the method could lead to skipping
an extra location in the case where the algorithm needed to generate
code for possibly applying an injection compaction. We change the
algorithm for skipping locations, so that we skip a location for every
child of an injection term.