Updating `imp.md`, kore tests and `kore-proof-trace` to parse side condition events (#1032)
This PR fixes https://github.com/Pi-Squared-Inc/pi2/issues/1348. This
issue reports that `kore-proof-trace` doesn't generate any hints when a
`requires` clause is used in an example.

After investigating the issue, we realized that our regression tests
didn't cover this case. Therefore, we updated the current `imp-*.kore`
tests to contain it. We also updated `imp.md` to be compatible with K
v7.0.15. To improve the reproducibility of our tests in the future, we
made `imp.md` the source code of `imp.kore`, `imp-proof.kore`, and
`imp-slow-proof.kore`.

Then, we needed to update the Proof Hints version and the
`kore-proof-trace` tool to include "side condition" events as arguments
and parse them appropriately to hints for the `requires` clause and the
rest of the test.