Initial commit of version scripts (#1033)
This PR drops in a copy of our usual version-bump scripts (previously
used in Pyk, currently in all semantics projects). We want this
versioning information to make CI improvements to K a bit more
ergonomic.

It explicitly does not run the scripts in CI yet, because we rely on
`package/version` existing in the PR base branch. I will follow this PR
up with a matching one that correctly bumps the version on PRs.