Bump versions and cut releases (#1035)
This is a follow-up PR to #1034 that implements version bumping on each
PR, and propagates that version information into places in the build
system where it's required.

Additionally, it adds a CI job on release that tags a GitHub release
with the appropriate metadata. Eventually, we will be attaching a `.deb`
build to that release, but for now it's just a tag.

---------

Co-authored-by: devops <devops@runtimeverification.com>