Build debian package (#1036)
This PR sets up a Debian package build for the LLVM backend, such that
we will be able to adapt the K CI testing process to download this
released version rather than rebuilding the backend from scratch on
every run.

The package building step is tested in CI, and I have tested the part of
the workflow that deals with _uploading_ the released package locally by
creating and deleting releases on pushes to this branch. I'll make sure
to babysit the release to ensure that the file is attached correctly
when this PR is merged.

The Debian packaging specifics in this PR have been largely adapted from
the main K repo.

---------

Co-authored-by: devops <devops@runtimeverification.com>