Fix usage of github PR SHA in release job (#1041)
This was an oversight when refactoring the release workflow code last
week; it's not easy to test this step without creating spurious PRs on
the K repo so it got missed. It being missed produced a
[failure](https://github.com/runtimeverification/k/actions/runs/8943558614/job/24594103312?pr=4312)
in the LLVM backend update job.

The fix is to send the correct version information from the stashed
version file to the devops job, rather than the Git SHA as we were
previously doing.

---------

Co-authored-by: devops <devops@runtimeverification.com>