Improve test harness and test coverage for proof hints (#1046)
Previously the tests for proof hints were rather anemic, not to mention
somewhat non-standard. We make the following changes in order to
canonify a technique for testing proof hints, and also expand the test
coverage of this code path:

1. We remove the kore-proof-hint-test binary which appears to be written
with the intent of being used in one specific integration test. This
seems to be duplicated logic with the test_proof_hint.py python tests,
and I don't see the need to maintain a specific C++ binary just for one
test. We also remove the test associated with this binary since it
appears to be redundant.
2. We clean up the infrastructure for how we test proof hints. We no
longer test proof hints without passing --expand-terms, since this
exposes information that is not particularly meaningful and makes the
tests flaky, like the number of bytes in each kore term in the trace,
which may be subject to change.
3. We codify a set of techniques for testing proof hints. Any test which
uses `%interpreter` and `%check-diff` can be extended by adding the
lines `%proof-interpreter` and `%check-proof-out` to the exact same kore
definition. This allows us to reuse kore files and input files and test
the same tests both with and without the proof hint.
4. We add the proof hint testing harness to a small selection of more
complex tests in the tests directory. If regressions occur in the
future, we can extend the number of tests we do this with, and we can
selectively choose to enable this test coverage on new integration tests
that we create very easily. We do not enable this on all tests because
it does not scale well to long execution traces due to the fact that
this would require very large expected output files to be stored in
source control.

---------

Co-authored-by: rv-jenkins <admin@runtimeverification.com>