[UnitTests] Adding `subsortMap` compiler unit tests (#1050)
This PR is part of investigating the `get_subsorts` function to ensure
it's working as expected.

This PR adds two small tests, the first just checking insertion,
contains, emptiness, and size of `subsortMap`. The second test builds a
`kore_defintion` as expected by `get_subsorts`, then calls the function
and checks if the `subsort2` is part of the `subsortSet` of `subsort1`.

---------

Co-authored-by: Bruce Collie <brucecollie82@gmail.com>