Remove injections from output files of proof hint tests (#1048)
Since the Pi2 team that consumes the proof hint is discarding the
injections, and we have a minor issue in our upcoming serialization PR
that does not quite deserialize injections correctly, I figured the
expedient thing to do would be to test whether or not the updated
serialization format provides the same output modulo the injections in
the provided terms. We implement this by simply recursively visiting
each term as it is printed by `kore-proof-trace` and stripping out the
injections prior to printing it.

This will be used to show that the deserialized output does not change
modulo the injections in a follow-up PR.

---------

Co-authored-by: rv-jenkins <admin@runtimeverification.com>