Serialize header for new serialization format (#1028)
This PR adds a tool to serialize the header for the new serialization
format that will be used by the proof hints. The new format focuses on
fast serialization of llvm backend terms rather than generality. It
should be quite fast to deserialize as well, however.

Not ready for review yet. I still have to:

* [x] write tests to ensure that the header emitted matches expectations
* [x] double check that this code doesn't break anything
* [x] write up documentation of the new serialization format