Don't emit branch unnecessary check for injection  (#1029)
Fix https://github.com/Pi-Squared-Inc/pi2/issues/1323
The llvm backend currently emits a branch each time a new injection is
created to check whether the term that is its child is itself an
injection. We skip over some circumstances where such a check would not
be necessary, but not all.
If the sort being injected has no subsort, then the child cannot be an
injection. This PR adds a condition to check if the current sort has no
subsorts. If that's the case, we also skip the branch/new blocks emit.

---------

Co-authored-by: devops <devops@runtimeverification.com>
Co-authored-by: rv-jenkins <admin@runtimeverification.com>