Implement serialization/deserialization of new algorithm (#1051)
This PR implements a new version of the proof hint, version 9, which
uses the newly created serialization and deserialization functions to
serialize terms during the proof hint generation. Not only does this new
serialization format take up less memory, it also is roughly 3x faster
to serialize.

Note: there is one known issue with this serializer/deserializer: it
incorrectly specifies the second sort parameter of the `inj` symbol when
deserializing. We choose to ignore this known issue due to the fact that
this deserializer is being consumed solely by code which transforms the
resulting terms into an order-sorted representation by removing the
injections entirely.