[Debugger] Introducing lldb on for `llvm-krun` script (#857)
Fixes #852 

This simple PR checks if the user is on a `macOS` system before
executing the interpreter in debug mode, and if that's the case, the
debug command is now modified to call `lldb` with the corrected
supported flags instead of calling `gdb`.

It also fixes a bug on `--debug-batch`.

---------

Co-authored-by: rv-jenkins <admin@runtimeverification.com>