Introducing `Ordinal` KORE Attribute (#1053)
Closes https://github.com/Pi-Squared-Inc/pi2/issues/1432
This PR introduces a new attribute specific to the LLVM Backend. It
includes the `ordinal` attribute to the KORE definition when
`preprocess` is called after parsing a definition. The content of the
attribute is a `string_sort` with the value of the computed
`axiom_ordinal`.

A new C++ tool is planned to consume this feature instead of reading a
definition with a bash script.
The main intention of this feature is to expose this information to the
Python bindings to be used in Proof Hints integration tests.