Introducing llvm-kompile `compute-ordinal` and `compute-loc` tools (#1056)
These tools are meant to replace the old `llvm-kompile-compute-ordinal`
and `llvm-kompile-compute-loc` bash scripts with an appropriate C++
infrastructure.

The new tools can work in two ways:
- Parsing the definition and looking for/using the K file location to
compute the ordinal and location info.
- Using the `definition.kore` lines to compute the ordinal and returning
the line in the definition when computing the location.

The interface of the new tools is defined as follows:
```bash
llvm-kompile-compute-loc [options] <kompiled-dir> <ordinal> [is-k-line]
```
and
```bash
llvm-kompile-compute-ordinal [options] <kompiled-dir> <line> [is-k-line]
```

in both cases, `is-k-line` is false by default to have the same behavior
as the old scripts.

---------

Co-authored-by: rv-jenkins <admin@runtimeverification.com>