Refactor proof hint parser (#1065)
This is an intermediate step on the way to having a streaming parser for
the proof hint. What this is doing is modifying the existing parser so
that it no longer loads the entire file into memory prior to parsing it,
and so that it abstracts away all the low-level I/O details behind an
interface.

This interface is an abstract class that is implemented by classes
representing in-memory and file buffers, respectively. In the future we
may also have a class representing a memory-mapped file.

In the next PR, I will create a new class that parses and returns events
one at a time.

---------

Co-authored-by: rv-jenkins <admin@runtimeverification.com>