Remove sentinel for end of KORE terms in the proof hint format (#1076)
This PR removes the sentinel at the end of a KORE terms as it would
appear in the proof hint trace. We do not need this sentinel to properly
parse the trace because the KORE term binary format includes length
information.

The hints version has been bumped to 11 and the relevant tests have been
updated.