Add python bindings for proof hint streaming parser (#1075)
Previously we did not have any python bindings for the streaming parser
I created.

This PR adds the relevant pybind code as well as adding a python unit
test for that code. A slight change is made to the ownership semantics
of proof_trace_buffer and its derived classes in order to allow the
python bindings to work as expected.

---------

Co-authored-by: rv-jenkins <admin@runtimeverification.com>