add flag to control whether gc should be enabled or not (#1082)
This is one of a sequence of PRs designed to make progress towards
generating stack maps so that we can trigger the GC during allocation
rather than in between rewrite steps only. The first few PRs will be
preliminaries that add small features that will be used by future PRs.

This PR adds a global variable that controls whether or not garbage
collection is enabled. This boolean flag is disabled during certain call
frames that it would be difficult to walk the stack of. In particular,
we disable the flag during:

* allocations that occur inside immer
* calls to hooks
* allocations that occur inside GMP and MPFR
* construction of the initial configuration

The boolean flag is currently unused. It will be used in a future pull
request to determine whether the allocator is allowed to trigger garbage
collection at the current time.

Co-authored-by: rv-jenkins <admin@runtimeverification.com>