create conditional compilation flag for stack maps (#1088)
This is one of a sequence of PRs designed to make progress towards
generating stack maps so that we can trigger the GC during allocation
rather than in between rewrite steps only. The first few PRs will be
preliminaries that add small features that will be used by future PRs.

In this PR we are creating a cmake compilation flag that enables or
disables the new gc strategy. It is disabled by default. Right now this
flag is unused, but it will be used in future PRs to guard the
in-progress functionality behind a flag so that it doesn't break
anything. In the future, once the feature is ready for testing, we will
make it enabled by default, and eventually delete the old behavior.

The cmake flag can then be accessed both by cmake and as a preprocessor
directive or boolean constant in C++ code.

Co-authored-by: rv-jenkins <admin@runtimeverification.com>