Update name of `EventType` and expose `__repr__` from `llvm_rewrite_trace_iterator` (#1090)
This small PR only updates the `LLVMEventType` name to `EventType` to
preserve the `LLVM` prefix to be used by the wrapper classes created on
PyK that export these bindings.

This PR also exposes the `__repr__` function of
`llvm_rewrite_trace_iterator`. The print function was already
implemented but has yet to be exposed; we decided to expose it as the
binding as it can be helpful for debugging.