add libunwind as dependency (#1081)
This is one of a sequence of PRs designed to make progress towards
generating stack maps so that we can trigger the GC during allocation
rather than in between rewrite steps only. The first few PRs will be
preliminaries that add small features that will be used by future PRs.

This PR adds the libunwind library as a runtime dependency to the llvm
backend. Libunwind is going to be used by the garbage collector to walk
the stack and find the GC roots for each stack frame. The code to
actually generate the stack maps and walk the stack will be added in a
future PR.

---------

Co-authored-by: rv-jenkins <admin@runtimeverification.com>